Keboola Dropbox Extractor (v2)
This commit contains a major overhaul in a way how the data downloaded from Dropbox are processed and uploaded into Keboola storage. The major thing is that the extractor assumes the user already authorized the links in GUI and the helpers extracted the configuration for that. The extractor just reads the CSV files and process them via streams without any modification, which saves a lot of memory consumption. The extractor also prepares particular manifest files.